Extract timer implementations into self-registering sub-packages
- timer/localtimer/ and timer/webtimer/ are now independent packages
  that register themselves via init() using timer.Register(Factory)
- timer package no longer imports its implementations; dependency arrows
  are reversed: implementations depend on the timer package, not vice versa
- timer.go (main) blank-imports both packages to trigger registration
- mob.go uses localtimer.ExecuteCommandsInBackgroundProcess/VoiceCommand
  directly for the moo feature
